add component main cities location
