+++ vite.config.ts
@@ -3,6 +3,7 @@
 
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
+import { resolve } from 'path';
 
 
 // https://vitejs.dev/config/
@@ -13,7 +14,6 @@ export default defineConfig({
     environment: 'jsdom',
     setupFiles: ['./src/setupTests.ts'],
   },
-
   resolve: {
     alias: [
       { find: '@', replacement: resolve(__dirname, 'src')
